+++ api/models/TaskModel.php
@@ -11,11 +11,12 @@ class TaskModel extends BaseModel
 
     public function add($data)
     {
-        $sql = "INSERT INTO `tasks` (title, category_id) VALUES (:title, :category_id)";
+        $sql = "INSERT INTO `tasks` (title, category_id, user_id) VALUES (:title, :category_id, :user_id)";
         $stmt = $this->db->prepare($sql);
         $stmt->execute([
             'title' => $data['title'],
             'category_id' => $data['category_id'],
+            'user_id' => $this->user_id,
         ]);
 
         $id = (int)$this->db->lastInsertId();
@@ -27,28 +28,31 @@ class TaskModel extends BaseModel
 
     public function edit($id, $data)
     {
-        $sql = "UPDATE `tasks` SET `title`= :title, `category_id`= :category_id, `completed` = :completed WHERE id = :id";
+        $sql = "UPDATE `tasks` SET `title`= :title, `category_id`= :category_id, `completed` = :completed WHERE id = :id and user_id = :user_id";
         $stmt = $this->db->prepare($sql);
         $stmt->execute([
             'id' => $id,
             'title' => $data['title'],
             'category_id' => $data['category_id'],
             'completed' => $data['completed'],
+            'user_id' => $this->user_id,
         ]);
         return [
             'id' => (int)$id,
             'title' => $data['title'],
             'category_id' => $data['category_id'],
             'completed' => $data['completed'],
+            'user_id' => $this->user_id,
         ];
     }
 
     public function deleteByCategoryId($id)
     {
-        $sql = "DELETE FROM `tasks` WHERE category_id = :category_id";
+        $sql = "DELETE FROM `tasks` WHERE category_id = :category_id and user_id = :user_id";
         $stmt = $this->db->prepare($sql);
         $stmt->execute([
             'category_id' => $id,
+            'user_id' => $this->user_id,
         ]);
         return [
             'id' => (int)$id
@@ -57,10 +61,11 @@ class TaskModel extends BaseModel
 
     public function deleteCompletedTasks($id)
     {
-        $sql = "DELETE FROM `tasks` WHERE category_id = :category_id and completed = 1";
+        $sql = "DELETE FROM `tasks` WHERE category_id = :category_id and completed = 1 and user_id = :user_id";
         $stmt = $this->db->prepare($sql);
         $stmt->execute([
             'category_id' => $id,
+            'user_id' => $this->user_id,
         ]);
         return [
             'id' => (int)$id
